archive/legacy/reconx_enterprise/reconx-enterprise/backend/requirements.txt
# Core Framework & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Async & Background Tasks
celery==5.3.4
python-dateutil==2.8.2

# HTTP & APIs
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Security & Auth
pyjwt==2.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Data Processing
pandas==2.1.4
numpy==1.24.3
networkx==3.2.1

# Parsing & Serialization
lxml==4.9.3
xmltodict==12.0.1
beautifulsoup4==4.12.2

# Code Quality & Testing (runtime optional)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
pytest-mock==3.12.0
black==23.12.1
flake8==6.1.0
pylint==3.0.3
isort==5.13.2
bandit==1.7.5
pre-commit==3.6.0

# Logging & Monitoring
python-json-logger==2.0.7
structlog==24.1.0

# CLI
typer==0.9.0
rich==13.7.0

# Visualization (optional)
matplotlib==3.8.2
plotly==5.18.0

# Machine Learning (optional - for ML models)
scikit-learn==1.3.2
joblib==1.3.2

# Rate limiting
slowapi==0.1.9

# Real-time updates
sse-starlette==2.0.0

# Pagination
fastapi-pagination==0.12.19

# CORS & Headers
python-cors==0.0.6

# Task scheduling
APScheduler==3.10.4

# UUID
shortuuid==1.0.13
